Added 6 test cases.
We created 3 test files for the 3 stalls we will have in our flea
market, each file will test the existence of the stall by checking the
location of the stall's sign. The file also tests the dialogues of each
NPC we plan to create in a separate file, name StallWorker 1, 2 and 3.