Implemented 3 new classes.
We created a class for a the Stall worker NPC, we used Wanda's class as
a template and edited it to suit our requirements. We created specific
dialogues for the NPC's in the flea market, and implemented them in a
way where they stand still without moving around. We also wrote the shop
sign's code in the same file as the NPC's, to make binding them easier
as we did. Finally we created each one of them with a random appearance.